fix controller, service, model, repository, exceptions

+++ src/main/java/com/wallet/api/walletapp/entity/Wallet.java
@@ -4,7 +4,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import lombok.Data;
 
-import java.math.BigDecimal;
 import java.util.UUID;
 
 @Entity
@@ -12,5 +11,5 @@ import java.util.UUID;
 public class Wallet {
     @Id
     private UUID id;
-    private BigDecimal balance;
+    private double balance;
 }
